Reorganize models into separate files for better structure
- Split monolithic models/student.py into organized files:
  - models/user.py: User and UserRole (authentication)
  - models/student.py: Student model only
  - models/professor.py: Professor model only
  - models/course.py: Course model and enrollment associations
- Update models/__init__.py to import from all new files
- Fix all import statements across backend and tests
- Update unit tests to work with new model structure
- Verify all functionality still works correctly
- All 23 model tests passing after reorganization
