+++ backend/config/auth.py
@@ -9,7 +9,7 @@ from fastapi import HTTPException, status, Depends
 from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
 from sqlalchemy.orm import Session
 from config.database import get_db
-from models.student import User, Student, Professor
+from models import User, Student, Professor
 from schemas.student_schemas import TokenData
 
 # Security configuration
