+++ backend/models/user.py
@@ -1,0 +1,25 @@
+"""
+User authentication models
+"""
+from sqlalchemy import Column, Integer, String, DateTime, Boolean
+from sqlalchemy.sql import func
+from config.database import Base
+import enum
+from sqlalchemy import Enum
+
+class UserRole(enum.Enum):
+    """User role enumeration"""
+    STUDENT = "student"
+    PROFESSOR = "professor"
+
+class User(Base):
+    """Base User model for authentication"""
+    __tablename__ = "users"
+    
+    id = Column(Integer, primary_key=True, index=True)
+    email = Column(String(255), unique=True, index=True, nullable=False)
+    hashed_password = Column(String(255), nullable=False)
+    role = Column(Enum(UserRole), nullable=False)
+    is_active = Column(Boolean, default=True)
+    created_at = Column(DateTime, default=func.now())
+    updated_at = Column(DateTime, default=func.now(), onupdate=func.now())
